requirements-dev.txt
# Neodoo18Framework - Odoo 18+ Development Dependencies
# Generated automatically by env.sh setup

# Core Odoo Dependencies
psycopg2-binary>=2.9.0
babel>=2.9.1
chardet>=4.0.0
cryptography>=3.4.8
decorator>=4.4.2
docutils>=0.17
ebaysdk>=2.1.5
feedparser>=6.0.8
freezegun>=1.2.0
gevent>=21.8.0
greenlet>=1.1.2
idna>=3.2
jinja2>=3.0.2
libsass>=0.21.0
lxml>=4.6.3
markupsafe>=2.0.1
num2words>=0.5.10
ofxparse>=0.21
passlib>=1.7.4
pillow>=9.0.0
polib>=1.1.1
psutil>=5.8.0
pydot>=1.4.2
pypdf2>=1.27.5
python-dateutil>=2.8.2
python-stdnum>=1.17
pytz>=2021.3
pyusb>=1.2.1
qrcode>=7.3.1
reportlab>=3.6.2
requests>=2.25.1
urllib3>=1.26.7
vobject>=0.9.6.1
werkzeug>=2.0.2
xlrd>=2.0.1
xlsxwriter>=3.0.1
xlwt>=1.3.0
zeep>=4.1.0

# Development Tools
pytest>=7.0.0
pytest-cov>=4.0.0
flake8>=5.0.0
black>=22.0.0
isort>=5.10.0
pre-commit>=2.20.0

# Framework-specific tools
click>=8.0.0
colorama>=0.4.0
rich>=12.0.0
PyYAML>=6.0
